+++ CHANGELOG.md
@@ -6,6 +6,7 @@ CHANGELOG
 
  * [BC BREAK] Rename service ID `ai.agent.response_format_factory` to `ai.platform.response_format_factory`
  * The `collection` option for `ChromaDb` is now optional.
+ * Update `DataCollector` to use `getCalls()` and `getResultCache()` getter methods on Traceable* classes
 
 0.7
 ---